project/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-jwt-extended==4.6.0
sqlalchemy==2.0.35
alembic==1.13.2
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1